--- frontend/src/pages/RoadTripPage/ItineraryPage/ItineraryPage.module.css
@@ -1,16 +1,0 @@
-.itineraryPage {
-    justify-content: center;
-    align-items: center;
-}
-
-.itineraryPage p {
-    color: #24305e;
-}
-
-.itineraryPageTitle {
-    font-size: 30px;
-}
-
-.emptyItineraryPageDescription {
-    font-style: italic;
-}
