+++ frontend/src/pages/RoadTripPageWrapper/RoadTripPage/ItineraryPage/ItineraryPage.jsx
@@ -1,6 +1,6 @@
 import { React } from 'react';
 import styles from './ItineraryPage.module.css';
-import ItineraryWrapper from '../../../components/ItineraryWrapper/ItineraryWrapper';
+import ItineraryWrapper from '../../../../components/ItineraryWrapper/ItineraryWrapper';
 
 export default function ItineraryPage() {
   return (
