+++ frontend/src/pages/RoadTripPageWrapper/RoadTripPage/RoadTripPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route, Redirect, useRouteMatch } from 'react-router-dom';
 
-import RoadTripTopBar from '../../components/RoadTripTopBar/RoadTripTopBar';
+import RoadTripTopBar from '../../../components/RoadTripTopBar/RoadTripTopBar';
 import MapPageWrapper from './MapPageWrapper/MapPageWrapper';
 import ItineraryPage from './ItineraryPage/ItineraryPage';
 import EmergencyDetailsPage from './EmergencyDetailsPage/EmergencyDetailsPage';
