+++ frontend/src/pages/RoadTripPageWrapper/RoadTripPage/SpotifyPlaylistPage/SpotifyPlaylistPage.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import useGet from '../../../hooks/useGet';
-import Spinner from '../../../components/commons/Spinner/Spinner';
-import SpotifyPlaylistWrapper from '../../../components/playlist/SpofityPlaylistPageWrapper/SpotifyPlaylistWrapper';
+import useGet from '../../../../hooks/useGet';
+import Spinner from '../../../../components/commons/Spinner/Spinner';
+import SpotifyPlaylistWrapper from '../../../../components/playlist/SpofityPlaylistPageWrapper/SpotifyPlaylistWrapper';
 
 export default function SpotifyPlaylistPage() {
   const { id } = useParams();
